--- a/cmake/tribits/common_tpls/FindTPLNetcdf.cmake
+++ b/cmake/tribits/common_tpls/FindTPLNetcdf.cmake
@@ -39,29 +39,29 @@
 
 #
 # First, set up the variables for the (backward-compatible) TriBITS way of
-# finding Netcdf.  These are used in case FIND_PACKAGE(NetCDF ...) is not
+# finding Netcdf.  These are used in case find_package(NetCDF ...) is not
 # called or does not find NetCDF.  Also, these variables need to be non-null
 # in order to trigger the right behavior in the function
-# TRIBITS_TPL_FIND_INCLUDE_DIRS_AND_LIBRARIES().
+# tribits_tpl_find_include_dirs_and_libraries().
 #
 
-SET(REQUIRED_HEADERS netcdf.h)
-SET(REQUIRED_LIBS_NAMES netcdf)
+set(REQUIRED_HEADERS netcdf.h)
+set(REQUIRED_LIBS_NAMES netcdf)
 
-IF (TPL_ENABLE_MPI)
-  SET(REQUIRED_LIBS_NAMES ${REQUIRED_LIBS_NAMES} pnetcdf)
-ENDIF()
+if (TPL_ENABLE_MPI)
+  set(REQUIRED_LIBS_NAMES ${REQUIRED_LIBS_NAMES} pnetcdf)
+endif()
 
 #
 # Second, search for Netcdf components (if allowed) using the standard
-# FIND_PACKAGE(NetCDF ...).
+# find_package(NetCDF ...).
 #
-TRIBITS_TPL_ALLOW_PRE_FIND_PACKAGE(Netcdf  Netcdf_ALLOW_PREFIND)
-IF (Netcdf_ALLOW_PREFIND)
+tribits_tpl_allow_pre_find_package(Netcdf  Netcdf_ALLOW_PREFIND)
+if (Netcdf_ALLOW_PREFIND)
 
-  MESSAGE("-- Using FIND_PACKAGE(Netcdf ...) ...")
+  message("-- Using find_package(Netcdf ...) ...")
 
-  SET(CMAKE_MODULE_PATH
+  set(CMAKE_MODULE_PATH
     "${CMAKE_MODULE_PATH}"
     "${CMAKE_CURRENT_LIST_DIR}/find_modules"
     "${CMAKE_CURRENT_LIST_DIR}/utils"
@@ -69,8 +69,8 @@ IF (Netcdf_ALLOW_PREFIND)
   
   find_package(NetCDF)
 
-  IF (NetCDF_FOUND)
-    SET(DOCSTR "List of semi-colon separated paths to look for the TPL Netcdf")
+  if (NetCDF_FOUND)
+    set(DOCSTR "List of semi-colon separated paths to look for the TPL Netcdf")
     set(TPL_Netcdf_Enables_Netcdf4 ${NetCDF_NEEDS_HDF5} CACHE BOOL
       "True if netcdf enables netcdf-4")
     set(TPL_Netcdf_Enables_PNetcdf ${NetCDF_NEEDS_PNetCDF} CACHE BOOL
@@ -83,32 +83,32 @@ IF (Netcdf_ALLOW_PREFIND)
       "List of semi-colon seprated library names (not 'lib' or extension).")
     set(TPL_Netcdf_INCLUDE_DIRS ${NetCDF_INCLUDE_DIRS} CACHE PATH
       "${DOCSTR} header files.")
-  ENDIF()
-ELSE()
+  endif()
+else()
   # Curl library is only required if DAP is enabled; should detect inside
   # FindNetCDF.cmake, but that is not being called... SEMS has DAP enabled;
   # many HPC systems don't, but they override these settings...
-  FIND_PROGRAM(NC_CONFIG "nc-config")
-  IF (NC_CONFIG)
-    EXECUTE_PROCESS(COMMAND "nc-config --has-dap2"
+  find_program(NC_CONFIG "nc-config")
+  if (NC_CONFIG)
+    execute_process(COMMAND "nc-config --has-dap2"
                     OUTPUT_VARIABLE NETCDF_HAS_DAP2)
-    EXECUTE_PROCESS(COMMAND "nc-config --has-dap4"
+    execute_process(COMMAND "nc-config --has-dap4"
                     OUTPUT_VARIABLE NETCDF_HAS_DAP4)
-  ENDIF()
-  IF ((NOT NC_CONFIG) OR NETCDF_HAS_DAP2 OR NETCDF_HAS_DAP4)
-    SET(REQUIRED_LIBS_NAMES ${REQUIRED_LIBS_NAMES} curl)
-  ENDIF()
-ENDIF()
+  endif()
+  if ((NOT NC_CONFIG) OR NETCDF_HAS_DAP2 OR NETCDF_HAS_DAP4)
+    set(REQUIRED_LIBS_NAMES ${REQUIRED_LIBS_NAMES} curl)
+  endif()
+endif()
 
 #
-# Third, call TRIBITS_TPL_FIND_INCLUDE_DIRS_AND_LIBRARIES()
+# Third, call tribits_tpl_find_include_dirs_and_libraries()
 #
-TRIBITS_TPL_FIND_INCLUDE_DIRS_AND_LIBRARIES( Netcdf
+tribits_tpl_find_include_dirs_and_libraries( Netcdf
   REQUIRED_HEADERS ${REQUIRED_HEADERS}
   REQUIRED_LIBS_NAMES ${REQUIRED_LIBS_NAMES}
   )
-# NOTE: If FIND_PACKAGE(Netcdf ...) was called and successfully found Netcdf,
-# then TRIBITS_TPL_FIND_INCLUDE_DIRS_AND_LIBRARIES() will use the already-set
+# NOTE: If find_package(Netcdf ...) was called and successfully found Netcdf,
+# then tribits_tpl_find_include_dirs_and_libraries() will use the already-set
 # variables TPL_Netcdf_INCLUDE_DIRS and TPL_Netcdf_LIBRARIES and then print
 # them out (and set some other standard variables as well).  This is the final
 # "hook" into the TriBITS TPL system.
@@ -117,7 +117,7 @@ TRIBITS_TPL_FIND_INCLUDE_DIRS_AND_LIBRARIES( Netcdf
 # Need to determine how this is set in the library that is being used...
 
 if ("${TPL_Netcdf_PARALLEL}" STREQUAL "")  
-   ASSERT_DEFINED(TPL_Netcdf_INCLUDE_DIRS)
+   assert_defined(TPL_Netcdf_INCLUDE_DIRS)
    find_path(meta_path
       NAMES "netcdf_meta.h"
       HINTS ${TPL_Netcdf_INCLUDE_DIRS}