This PR contains a few changes to Trilinos 'develop' needed to work with updated TriBITS or updated CMake (even without TriBITS changes).

See the individual commits for more details which include the following changes:

* Fix CMake build error related to math library spaces from ATDM Trilinos configuration
* Fix non-TriBITS `add_exectuable()` code in Sacado to work with old and new TriBITS
* Fix Zoltan2 tests that were using explicit semi-colons (which is understandable because that is what the documentation at the time said and it worked at the time.  See https://github.com/TriBITSPub/TriBITS/pull/481).  (This changing logic added 9 years ago in commit c9f5e7bd073901eee3b69cded20e3dcc989f60e1.)
* Unconditionally disable Isorropia EpetraExt tests that have not been built in over 10 years (see #10534)
* Make CMake function/macro calls lower-case for FindTPLBLAS.cmake, FindTPLLAPACK.cmake, and FindTPLNetdf.cmake.  (These three files have diffs compared to TriBITS 'master' so I wanted to lower-case the function/macro calls so that I can see the real changes by diffing the raw files.)
 
Now that as part of the SEMS PR clang-7.0.1 build below, I also did some configure timings for enabling all of the Trilinos packages and we are seeing a 25% overall speedup in configure+generate times for the new TriBITS vs. the old TriBITS.  The results were:

* Old TriBITS: configure = 35.1s, configure+generate = 2m25s, generate = 110s
* New TriBiTS: configure = 28.3s, configure+generate = 1m48s, generate = 80s

The details are given below.

<details>

<summary><b>Details for configure timing for SEMS PR clang-7.0.1 build</b> (click to expand)</summary>

<br>

This is for the repos versions:

```
$ gitdist-status --dist-repos=.,TriBITS
-----------------------------------------------------------------------------------------------------------------------------------
| ID | Repo Dir        | Branch                                  | Tracking Branch                                    | C | M | ? |
|----|-----------------|-----------------------------------------|----------------------------------------------------|---|---|---|
|  0 | Trilinos (Base) | tribits-299-modern-cmake-targets-2-prep | rab-github/tribits-299-modern-cmake-targets-2-prep |   |   |   |
|  1 | TriBITS         | 299-switch-to-modern-cmake-targets      | rab-github/299-switch-to-modern-cmake-targets      |   |   |   |
-----------------------------------------------------------------------------------------------------------------------------------

$ gitdist-repo-versions --dist-repos=.,TriBITS
*** Base Git Repo: Trilinos
8c1028df724eb0096cd688b381285112bc6f914a [Wed May 18 15:08:17 2022 -0600] <rabartl@sandia.gov>
FindTPLNetcdf.cmake: Lower-case function names (TriBITSPub/TriBITS#274)
*** Git Repo: TriBITS
97f8d6c03ebadc46dfeaac2a4d9c555c3831f590 [Wed May 18 16:53:37 2022 -0600] <rabartl@sandia.gov>
Snapshot FindTPLNetcdf.cmake from Trilinos 8c1028df724 with small mods (#299)
```

What is interesting is that the reported configure times are:

* Old TriBITS: configure = 35.1s, configure+generate = 2m25s
* New TriBiTS: configure = 28.3s, configure+generate = 1m48s

I wonder if local timings would show this out?

First the reference build with old TriBITS:

```
$ cd /ascldap/users/rabartl/Trilinos.base/BUILDS/PR/clang-10.0.0.ref/

$ . load-env.sh

$ for i in 0 1 ; do time cmake . &> configure.out ; done ; grep -B 1 "Total time to configure Trilinos" configure.out

real    2m19.760s
user    1m13.621s
sys     0m43.884s

real    2m20.824s
user    1m14.111s
sys     0m43.836s

Total time to configure Trilinos: 0m31.715s
```

Then the updated build with new TriBITS:

```
$ cd /ascldap/users/rabartl/Trilinos.base/BUILDS/PR/clang-10.0.0.tribits/

$ . load-env.sh

$ for i in 0 1 ; do time cmake . &> configure.out ; done ; grep -B 1 "Total time to configure Trilinos" configure.out

real    1m47.398s
user    0m56.726s
sys     0m41.518s

real    1m46.337s
user    0m56.577s
sys     0m39.812s

Total time to configure Trilinos: 0m27.431s
```

So this reduction in configure and especially generate times with the new TriBITS implementation are real!  But the larger real reduction is in the generation times:

* Old TriBITS: configure = 35.1s, configure+generate = 2m25s, generate = 110s
* New TriBiTS: configure = 28.3s, configure+generate = 1m48s, generate = 80s

It makes sense that the new TriBITS implementation configure step would be faster but I don't know why the generate step would be so much faster.

That is a reduction of 25%.  That is a pretty significant reduction in configure+generate times!

</details>


## Testing

### ATDM Trilinos clang 7.0.1 build 

I developed these fixes and made these changes to get the Trilinos test suite to pass with the ATDM Trilinos build configuration `sems-rhel7-clang-7.0.1-openmp-shared-release-debug` which I modified to use updated TriBITS.  The testing results were:

```
100% tests passed, 0 tests failed out of 2250
```

<details>

<summary><b>sems-rhel7-clang-7.0.1-openmp-shared-release-debug testing details</b> (click to expand)</summary>

.

Here are the repo state and versions on crf450:

```
$ cd /home/rabartl/Trilinos.base/Trilinos/

$ gitdist-status --dist-repos=.,TriBITS
-----------------------------------------------------------------------------------------------------------------------------------
| ID | Repo Dir        | Branch                                  | Tracking Branch                                    | C | M | ? |
|----|-----------------|-----------------------------------------|----------------------------------------------------|---|---|---|
|  0 | Trilinos (Base) | tribits-299-modern-cmake-targets-2-prep | rab-github/tribits-299-modern-cmake-targets-2-prep |   |   |   |
|  1 | TriBITS         | 299-switch-to-modern-cmake-targets      | rab-github/299-switch-to-modern-cmake-targets      |   |   |   |
-----------------------------------------------------------------------------------------------------------------------------------

$ gitdist-repo-versions --dist-repos=.,TriBITS
*** Base Git Repo: Trilinos
82e640872372d9784a846921e1a5ace14ecb074d [Mon May 16 10:47:53 2022 -0600] <rabartl@sandia.gov>
Zoltan2: Use single value for FAIL_REGULAR_EXPRESSION (TriBITSPub/TriBITS#464)
*** Git Repo: TriBITS
29d741095b3f174466edd5d2d777713921dea122 [Mon May 16 19:27:50 2022 -0600] <rabartl@sandia.gov>
TribitsExampleProject2: Remove hacked dependencies on indirect TPLs (#63, #299)
```

The env and driver scripts on crf450 are:

```
$ cd /home/rabartl/Trilinos.base/BUILDS/ATDM/SEMS-RHEL7/CHECKIN/sems-rhel7-clang-7.0.1-openmp-shared-release-debug

$ cat load-env.sh 
source /home/rabartl/Trilinos.base/Trilinos/cmake/std/atdm/load-env.sh sems-rhel7-clang-7.0.1-openmp-shared-release-debug

$ cat do-configure
./do-configure.base \
-DTrilinos_ENABLE_ALL_OPTIONAL_PACKAGES:BOOL=ON \
-DTrilinos_ENABLE_ALL_PACKAGES:BOOL=ON \
-DTrilinos_ENABLE_ALL_FORWARD_DEP_PACKAGES:BOOL=OFF \
-DTrilinos_ENABLE_Krino:BOOL=OFF \
"$@"

$ cat do-configure.base 
cmake \
-GNinja \
-DTrilinos_TRIBITS_DIR:STRING=TriBITS/tribits \
-DTrilinos_ENABLE_TESTS:BOOL=ON \
-DTrilinos_TEST_CATEGORIES:STRING=NIGHTLY \
-DTrilinos_ALLOW_NO_PACKAGES:BOOL=OFF \
-DDART_TESTING_TIMEOUT:STRING=600.0 \
-GNinja \
-DTrilinos_CONFIGURE_OPTIONS_FILE:STRING=cmake/std/atdm/ATDMDevEnv.cmake \
-DTrilinos_TRACE_ADD_TEST=ON \
"$@" \
/home/rabartl/Trilinos.base/Trilinos/cmake/std/atdm/../../..

```

Loading env and (re)building and testing:

```
$ cd /home/rabartl/Trilinos.base/BUILDS/ATDM/SEMS-RHEL7/CHECKIN/sems-rhel7-clang-7.0.1-openmp-shared-release-debug

$ . load-env.sh 
Hostname 'crf450' matches known ATDM host 'crf450' and system 'sems-rhel7'
Setting compiler and build options for build-name 'sems-rhel7-clang-7.0.1-openmp-shared-release-debug'
Using SEMS RHEL7 compiler stack CLANG-7.0.1 to build RELEASE-DEBUG code with Kokkos node type OPENMP

$ rm -r CMake* *.out

$ time ./do-configure &> configure.out

real    0m53.812s
user    0m32.776s
sys     0m18.785s

$ time make &> make.out

real    0m2.791s
user    0m22.691s
sys     0m5.629s

$ time ctest -j16  &> ctest.out

real    7m44.568s
user    147m47.744s
sys     14m57.467s

$ grep -A 500 "failed out of" ctest.out

100% tests passed, 0 tests failed out of 2250

Subproject Time Summary:
Adelus                     =   6.10 sec*proc (4 tests)
Amesos2                    =   4.72 sec*proc (6 tests)
Belos                      = 211.17 sec*proc (144 tests)
Ifpack2                    =  50.21 sec*proc (51 tests)
Intrepid2                  = 759.52 sec*proc (350 tests)
Kokkos                     = 493.44 sec*proc (45 tests)
KokkosKernels              = 260.41 sec*proc (23 tests)
MueLu                      = 537.52 sec*proc (108 tests)
NOX                        = 220.85 sec*proc (110 tests)
Panzer                     = 2912.14 sec*proc (196 tests)
Phalanx                    =  18.89 sec*proc (36 tests)
Rythmos                    = 230.44 sec*proc (83 tests)
SEACAS                     =  37.23 sec*proc (28 tests)
STK                        =   2.43 sec*proc (1 test)
Sacado                     = 123.52 sec*proc (305 tests)
Stratimikos                =  43.54 sec*proc (40 tests)
Teko                       = 101.38 sec*proc (18 tests)
Teuchos                    = 100.86 sec*proc (146 tests)
Thyra                      =  49.85 sec*proc (82 tests)
Tpetra                     = 288.60 sec*proc (273 tests)
TrilinosATDMConfigTests    =  54.78 sec*proc (12 tests)
TrilinosBuildStats         =   5.43 sec*proc (2 tests)
Xpetra                     =  15.57 sec*proc (19 tests)
Zoltan2                    = 279.94 sec*proc (168 tests)

Total Test time (real) = 464.48 sec
```

</details>

### SEMS PR clang-7.0.1 build

I (sort of) reproduced the SEMS-based clang-10.0.0 PR build locally on my RHEL7 machine 'crf450'.  After getting the TriBITS copy of FindTPLNetcdf.cmake to match what is in Trilinos 'develop' (see https://github.com/TriBITSPub/TriBITS/pull/479/commits/97f8d6c03ebadc46dfeaac2a4d9c555c3831f590), I was able to get all passing tests with both the old and new TriBITS (in https://github.com/TriBITSPub/TriBITS/pull/479).  I posted to CDash shown at:

* https://testing.sandia.gov/cdash/index.php?project=Trilinos&date=2022-05-19&filtercount=1&showfilters=1&field1=site&compare1=61&value1=crf450

which shows:

![image](https://user-images.githubusercontent.com/1959736/169575586-7138994b-a666-4855-94ff-ea62577ba9df.png)

The build Linux-clang-10.0.0.ref used the existing TriBITS under Trilihnos/cmake/tribits while the build Linux-clang-10.0.0.tribits usesd the new TriBITS in PR https://github.com/TriBITSPub/TriBITS/pull/479.

NOTE: These builds actually used clang-7.0.1 instead of clang-10.0.0 as I reported in #10540.  But the updated TriBITS build passed compared to the reference build so it does not really matter for this purpose.

More details are given below.

<details>

<summary><b>Details for SEMS clang-10.0.0 (clang-7.0.1) builds</b> (click to expand)</summary>

.

The repo versions for these tests were:

```
$ gitdist-status --dist-repos=.,TriBITS
-----------------------------------------------------------------------------------------------------------------------------------
| ID | Repo Dir        | Branch                                  | Tracking Branch                                    | C | M | ? |
|----|-----------------|-----------------------------------------|----------------------------------------------------|---|---|---|
|  0 | Trilinos (Base) | tribits-299-modern-cmake-targets-2-prep | rab-github/tribits-299-modern-cmake-targets-2-prep |   |   |   |
|  1 | TriBITS         | 299-switch-to-modern-cmake-targets      | rab-github/299-switch-to-modern-cmake-targets      |   |   |   |
-----------------------------------------------------------------------------------------------------------------------------------

$ gitdist-repo-versions --dist-repos=.,TriBITS
*** Base Git Repo: Trilinos
8c1028df724eb0096cd688b381285112bc6f914a [Wed May 18 15:08:17 2022 -0600] <rabartl@sandia.gov>
FindTPLNetcdf.cmake: Lower-case function names (TriBITSPub/TriBITS#274)
*** Git Repo: TriBITS
97f8d6c03ebadc46dfeaac2a4d9c555c3831f590 [Wed May 18 16:53:37 2022 -0600] <rabartl@sandia.gov>
Snapshot FindTPLNetcdf.cmake from Trilinos 8c1028df724 with small mods (#299)
```

Doing the updated TriBITS bulid:

```
$ ssh crf450

$ cd /ascldap/users/rabartl/Trilinos.base/BUILDS/PR/clang-10.0.0.tribits/

[rabartl@crf450 clang-10.0.0.tribits (crf450)]$ cat load-env.sh 
source /ascldap/users/rabartl/Trilinos.base/Trilinos/cmake/std/sems/PullRequestClang10.0.0TestingEnv.sh
module unload sems-archive-python

$ cat do-configure.base
TRILINOS_DIR=${HOME}/Trilinos.base/Trilinos
if [ -e CMakeCache.txt ] ; then
  rm -r CMake*
fi
cmake \
-C ${TRILINOS_DIR}/cmake/std/PullRequestLinuxClang10.0.0TestingSettings.cmake \
-G Ninja \
-D Trilinos_TRIBITS_DIR:STRING=TriBITS/tribits \
-D CTEST_BUILD_FLAGS=-j20 \
-D CTEST_PARALLEL_LEVEL=16 \
"$@" \
${TRILINOS_DIR}

$ cat do-configure
./do-configure.base \
-D Trilinos_ENABLE_TESTS=ON \
-D Trilinos_ENABLE_ALL_PACKAGES=ON \
"$@"

$ . load-env.sh

$ rm *.out

$ time ./do-configure &> configure.out

real    2m5.069s
user    1m10.377s
sys     0m50.419s

$ grep Trilinos_TRIBITS_DIR configure.out 
-- Trilinos_TRIBITS_DIR='/ascldap/users/rabartl/Trilinos.base/Trilinos/cmake/tribits'

$ grep -B 1 "Final set of" configure.out 

Final set of enabled packages:  TrilinosFrameworkTests TrilinosATDMConfigTests Gtest Kokkos Teuchos KokkosKernels RTOp Sacado MiniTensor Epetra Zoltan Shards Triutils EpetraExt Tpetra TrilinosSS Domi Thyra Xpetra Isorropia Pliris AztecOO Galeri Amesos Pamgen Zoltan2Core Ifpack ML Belos ShyLU_Node Amesos2 SEACAS Komplex Anasazi Ifpack2 Stratimikos FEI Teko Intrepid Intrepid2 Compadre STK Percept Krino Phalanx NOX Moertel MueLu Zoltan2Sphynx Zoltan2 ShyLU_DD ShyLU Rythmos Tempus Stokhos ROL Piro Panzer Adelus TrilinosCouplings Pike TrilinosBuildStats 62

Final set of enabled SE packages:  TrilinosFrameworkTests TrilinosATDMConfigTests Gtest KokkosCore KokkosContainers KokkosAlgorithms Kokkos TeuchosCore TeuchosParser TeuchosParameterList TeuchosComm TeuchosNumerics TeuchosRemainder TeuchosKokkosCompat TeuchosKokkosComm Teuchos KokkosKernels RTOp Sacado MiniTensor Epetra Zoltan Shards Triutils EpetraExt TpetraClassic TpetraTSQR TpetraCore Tpetra TrilinosSS Domi ThyraCore ThyraEpetraAdapters ThyraEpetraExtAdapters ThyraTpetraAdapters Thyra Xpetra Isorropia Pliris AztecOO Galeri Amesos Pamgen Zoltan2Core Ifpack ML Belos ShyLU_NodeHTS ShyLU_NodeTacho ShyLU_Node Amesos2 SEACASExodus SEACASExodus_for SEACASExoIIv2for32 SEACASNemesis SEACASIoss SEACASChaco SEACASAprepro_lib SEACASSupes SEACASSuplib SEACASSuplibC SEACASSuplibCpp SEACASAlgebra SEACASAprepro SEACASConjoin SEACASEjoin SEACASEpu SEACASCpup SEACASExodiff SEACASExomatlab SEACASExotxt SEACASExo_format SEACASEx1ex2v2 SEACASGjoin SEACASGen3D SEACASGenshell SEACASGrepos SEACASExplore SEACASMapvarlib SEACASMapvar SEACASMapvar-kd SEACASNas2exo SEACASZellij SEACASNemslice SEACASNemspread SEACASNumbers SEACASSlice SEACASTxtexo SEACASEx2ex1v2 SEACAS Komplex Anasazi Ifpack2 Stratimikos FEI Teko Intrepid Intrepid2 Compadre STKUtil STKEmend STKCoupling STKMath STKSimd STKNGP_TEST STKTopology STKMesh STKIO STKSearch STKTransfer STKTools STKBalance STKUnit_test_utils STKSearchUtil STKUnit_tests STKDoc_tests STKExprEval STK Percept Krino Phalanx NOX Moertel MueLu Zoltan2Sphynx Zoltan2 ShyLU_DDBDDC ShyLU_DDFROSch ShyLU_DDCommon ShyLU_DD ShyLU Rythmos Tempus Stokhos ROL Piro PanzerCore PanzerDofMgr PanzerDiscFE PanzerAdaptersSTK PanzerMiniEM Panzer Adelus TrilinosCouplings PikeBlackBox PikeImplicit Pike TrilinosBuildStats 148

Final set of non-enabled packages:  TriKota PyTrilinos NewPackage 3

Final set of non-enabled SE packages:  ShyLU_NodeBasker ShyLU_NodeFastILU SEACASSVDI SEACASPLT SEACASBlot SEACASExo2mat SEACASExotec2 SEACASFastq SEACASMat2exo TriKota ShyLU_DDCore PanzerExprEval PyTrilinos NewPackage 14

Final set of enabled TPLs:  Pthread MPI BLAS LAPACK Boost Scotch ParMETIS Zlib HDF5 Netcdf SuperLU BoostLib DLlib 13

Final set of non-enabled TPLs:  MKL yaml-cpp Peano CUDA CUBLAS CUSOLVER CUSPARSE Thrust Cusp TBB HWLOC QTHREAD BinUtils ARPREC QD OVIS gpcd METIS MTMETIS PuLP TopoManager LibTopoMap PaToH CppUnit ADOLC ADIC TVMET MF ExodusII Nemesis XDMF CGNS Pnetcdf ADIOS2 y12m SuperLUDist SuperLUMT Cholmod UMFPACK MA28 AMD CSparse HYPRE PETSC BLACS SCALAPACK MUMPS STRUMPACK PARDISO_MKL PARDISO Oski TAUCS ForUQTK Dakota HIPS MATLAB CASK SPARSKIT QT gtest BoostAlbLib OpenNURBS Portals CrayPortals Gemini InfiniBand BGPDCMF BGQPAMI Pablo HPCToolkit Clp GLPK qpOASES Matio PAPI MATLABLib Eigen X11 Lemon GLM quadmath CAMAL RTlib AmgX CGAL CGALCore VTune TASMANIAN ArrayFireCPU SimMesh SimModel SimParasolid SimAcis SimField Valgrind QUO ViennaCL Avatar mlpack pebbl MAGMASparse Check SARMA 103

$ time make &> make.out

real    0m26.508s
user    7m54.923s
sys     1m11.795s

$ time ctest -j16 &> ctest.out

real    11m37.899s
user    155m10.184s
sys     14m57.103s

$ grep "failed out of" ctest.out 
100% tests passed, 0 tests failed out of 3368
```

The reference build with the old TriBITS was identical except for setting

```
  -D Trilinos_TRIBITS_DIR:STRING=cmake/tribits
```

</details>